requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
exceptiongroup==1.3.0
fastmcp==2.9.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.10.1
mdurl==0.1.2
openapi-pydantic==0.5.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.35.0
